Moving bracket building to separate service file. p5js scss file. Pushing forward re-render of bracket
