+++ services/frontend/src/App.jsx
@@ -19,7 +19,7 @@ import { ThemeProvider } from './ThemeContext';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap';
 import DownloadableP5Image from './components/Bracket';
-import Testp5Sketch from './pages/TestP5Sketch';
+import ImageGeneratorWithoutWrapper from './components/DownloadImage';
 
 function App() {
   return (
@@ -36,7 +36,7 @@ function App() {
               <Route path="/bracket" element={<DownloadableP5Image />} />
               <Route path="/login" element={<LoginPage />} />
               <Route path="/create_account" element={<CreateAccountPage />} />
-              <Route path="/test-sketch" element={<Testp5Sketch />} />
+              <Route path="test_download" element={<ImageGeneratorWithoutWrapper />} />
               <Route path="*" element={<NotFoundPage />} />
             </Routes>
           </main>
