+++ services/frontend/src/App.scss
@@ -1,5 +1,6 @@
 @use 'stylesheets/custom.scss';
 @use 'stylesheets/progress.scss';
+@use 'stylesheets/p5js.scss';
 
 .App {
   text-align: center;
