+++ services/frontend/src/stylesheets/p5js.scss
@@ -1,0 +1,4 @@
+.bracket-canvas-continer > canvas {
+  max-width: 100%;
+  height: auto !important;
+}
